apache-superset-4.1.1rc1/api/requirements.txt
flasgger==0.9.7.1
mysql-connector-python
Flask==3.0.0
Flask_RESTful==0.3.10
flask_swagger_ui==4.11.1
flask-cors==4.0.0
pyodbc==5.2.0